Pull the extension method out of Renderer to avoid confusion.
